Support custom file names for collections
When adding a collection, the file argument can be passed which sets a
custom file name for the collection. By default, the file name is equal
to the collection name. Setting the same custom file name allows several
collections to be persisted to (and loaded from) the same JSON file.
This can be desirable for small collections to avoid having to make a
separate HTTP request for every single collection.