File/web paths: Store minified as .min.json
Instead of .json and -raw.json, use .min.json and .json for minified
and human-readably JSON files respectively. The previous approach is
susceptible to name conflicts when two collections exist, such as
"countries" and "countries-raw". Since periods are not permitted in
collection or file names, the new file/web paths do not have the issue
of potential file name conflicts.